requirements.txt
# Production Dependencies
openai>=1.0.0
typer>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0
GitPython>=3.1.0
Markdown>=3.4.0


# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Code Quality
ruff>=0.1.0


# Generated from: .venv/bin/pip freeze
# Resolved versions from virtual environment:
# typer==0.21.1
# rich==14.2.0
# pydantic==2.12.5
# pydantic-settings==2.12.0
# PyYAML==6.0.3
# GitPython==3.1.46
